feat: [BBND-1498] move Dividend domain models from equity to dividend context
