+++ packages/ats/sdk/__tests__/port/environmentMock.ts
@@ -7,7 +7,7 @@ import { SecurityRole } from "@domain/context/security/SecurityRole";
 import { EquityDetails } from "@domain/context/equity/EquityDetails";
 import { BondDetails } from "@domain/context/bond/BondDetails";
 import { CouponDetails } from "@domain/context/bond/CouponDetails";
-import { Dividend } from "@domain/context/equity/Dividend";
+import { Dividend } from "@domain/context/dividend/Dividend";
 import { VotingRights } from "@domain/context/equity/VotingRights";
 import { Coupon } from "@domain/context/bond/Coupon";
 import { ScheduledSnapshot } from "@domain/context/security/ScheduledSnapshot";
@@ -27,7 +27,7 @@ import Injectable from "@core/injectable/Injectable";
 import { CLIENT_PUBLIC_KEY_ECDSA } from "../config";
 import { CastRegulationSubType, CastRegulationType } from "@domain/context/factory/RegulationType";
 import { ScheduledBalanceAdjustment } from "@domain/context/equity/ScheduledBalanceAdjustment";
-import { DividendFor } from "@domain/context/equity/DividendFor";
+import { DividendFor } from "@domain/context/dividend/DividendFor";
 import { VotingFor } from "@domain/context/equity/VotingFor";
 import { Kyc } from "@domain/context/kyc/Kyc";
 import { KycAccountData } from "@domain/context/kyc/KycAccountData";
