+++ packages/ats/sdk/src/domain/context/dividend/DividendAmountFor.ts
